% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livestreamStart.R
\name{livestreamStart}
\alias{livestreamStart}
\title{Initiate a live coding stream using nGrok (https://ngrok.com/)}
\usage{
livestreamStart(hostname = NULL, username = NULL, password = NULL, port = 4040)
}
\arguments{
\item{hostname}{webpage (as character string) to host the live stream}

\item{username}{character string username used to access live stream}

\item{password}{character string password used to access live stream}

\item{port}{default is 4040. what web port to use (https://ngrok.com/docs/secure-tunnels/ngrok-agent/web-inspection-interface/)}
}
\value{
id value of tunnel that can be used to shut down the tunnel later
}
\description{
Initiate a live coding stream using nGrok (https://ngrok.com/)
}
\details{
NOTE: this package assumes that you have followed the steps to install nGrok on your computer (https://ngrok.com/docs/getting-started/ also https://bitsandbricks.github.io/post/code-live-from-rstudio-and-share-it-with-the-world-in-real-time/. For troubleshooting install, see https://stackoverflow.com/questions/65251887/clang-7-error-linker-command-failed-with-exit-code-1-for-macos-big-sur/65334247#65334247.
}
