% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livestreamStop.R
\name{livestreamStop}
\alias{livestreamStop}
\title{Stop a live coding stream using nGrok (https://ngrok.com/)}
\usage{
livestreamStop(id = NULL)
}
\arguments{
\item{id}{id of the tunnel instance}
}
\description{
Stop a live coding stream using nGrok (https://ngrok.com/)
}
