% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngrok.R
\name{ngrok}
\alias{ngrok}
\title{Install ngrok}
\usage{
ngrok(port = 8080, user = NULL, password = NULL)
}
\description{
\code{ngrok()} provide the function to start ngrok job.
\code{proc_kill()} used to kill the ngrok.
}
